requirements.txt
# Coastal Labor-Resilience Engine
# Core dependencies

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Scientific computing
scipy>=1.10.0

# HTTP requests for APIs
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Date/time handling
pytz>=2023.3

# Progress bars
tqdm>=4.65.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Data visualization (optional but useful)
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================================================
# Phase 2: NLP Classification
# ============================================================================

# OpenAI API for LLM classification
openai>=1.0.0

# Scikit-learn for evaluation metrics (optional)
scikit-learn>=1.3.0

# ============================================================================
# Phase 4: Frontend & Visualization
# ============================================================================

# Streamlit dashboard
streamlit>=1.28.0

# Interactive charts
plotly>=5.18.0

# Geospatial visualization (optional, for advanced maps)
# folium>=0.15.0
# pydeck>=0.8.0

